chore: refactor into smaller components
