+++ components/TaskListItem.tsx
@@ -1,0 +1,45 @@
+import ClayLabel from '@clayui/label';
+import ClayList from '@clayui/list';
+import Link from 'next/link';
+import React from 'react';
+
+import STATES from '../constants/taskStates';
+import formatDuration from '../utils/formatDuration';
+import CountUp from './CountUp';
+
+const TaskListItem = ({task, taskState}) => (
+	<ClayList.Item flex key={task.id}>
+		<ClayList.ItemField expand>
+			<Link href={`/tasks/${task.id}`} passHref>
+				<ClayList.ItemTitle>{task.name}</ClayList.ItemTitle>
+			</Link>
+
+			<ClayList.ItemText>
+				{task.totalRecommendations} Recommendations
+			</ClayList.ItemText>
+		</ClayList.ItemField>
+
+		{taskState === STATES.runningState && (
+			<ClayList.ItemField>
+				<CountUp startTime={task.startTime} />
+			</ClayList.ItemField>
+		)}
+
+		{taskState === STATES.completedState && (
+			<ClayList.ItemField>
+				{formatDuration(
+					new Date(task.finishedTime).getTime() -
+						new Date(task.startTime).getTime()
+				)}
+			</ClayList.ItemField>
+		)}
+
+		<ClayList.ItemField>
+			<ClayLabel displayType={STATES.byId[task.state].displayType}>
+				{STATES.byId[task.state].label}
+			</ClayLabel>
+		</ClayList.ItemField>
+	</ClayList.Item>
+);
+
+export default TaskListItem;
