+++ pages/index.tsx
@@ -1,9 +1,5 @@
 export default function Home() {
-	return (
-		<>
-			<h1>Dashboard</h1>
-		</>
-	);
+	return <></>;
 }
 
 export async function getServerSideProps() {
